ethers: Move files out of src/ethers directory
