--- ethers-ext/src/ethers/index.ts
@@ -1,2 +1,0 @@
-export { Wallet } from "./signer";
-export { Accounts, AccountStore } from "./accountStore";
