+++ ethers-ext/src/index.ts
@@ -1,9 +1,6 @@
-// TODO: move files to src/*.ts
-export { formatKlaytnUnits, formatKlay, parseKlaytnUnits, parseKlay } from "./core/util";
-export { Wallet, Accounts, AccountStore } from "./ethers";
-export { verifyMessageAsKlaytnAccountKey } from "./ethers/signer";
-
 export * from "@klaytn/js-ext-core/util";
 
-export * from "./provider";
+export * from "./accountStore";
 export * from "./keystore";
+export * from "./provider";
+export * from "./signer";
